Moved: initializer to `lib`
