+++ plugin.video.retrospect/resources/lib/__init__.py
@@ -1,6 +1,7 @@
 # SPDX-License-Identifier: CC-BY-NC-SA-4.0
 
 __all__ = ["addon", "addonsettings", "backtothefuture", "channelinfo", "chn_class", "cloaker",
-           "contextmenu", "envcontroller", "environments", "favourites", "locker", "logger",
-           "mediaitem", "menu", "paramparser", "parserdata", "pickler", "plugin", "proxyinfo",
-           "regexer", "retroconfig", "updater", "urihandler", "vault", "version", "xbmcwrapper"]
+           "contextmenu", "envcontroller", "environments", "favourites", "initializer", "locker",
+           "logger", "mediaitem", "menu", "paramparser", "parserdata", "pickler", "plugin",
+           "proxyinfo", "regexer", "retroconfig", "updater", "urihandler", "vault", "version",
+           "xbmcwrapper"]
