+++ plugin.video.retrospect/resources/lib/initializer.py
